+++ tests/unit/app/tSpectrogram.m
@@ -45,5 +45,89 @@ classdef tSpectrogram < MAGControllerTestCase
             testCase.verifyEqual(spectrogram.WindowSpinner.Layout, matlab.ui.layout.GridLayoutOptions(Row = 5, Column = [2, 3]), ...
                 "Window spinner layout should match expectation.");
         end
+
+        % Test that "getVisualizeCommand" returns expected command.
+        function getVisualizeCommand(testCase)
+
+            % Set up.
+            panel = testCase.createTestPanel();
+
+            spectrogram = mag.app.control.Spectrogram(panel);
+            spectrogram.instantiate();
+
+            results = mag.Instrument();
+
+            % Exercise.
+            command = spectrogram.getVisualizeCommand(results);
+
+            % Verify.
+            testCase.verifyEqual(command.PositionalArguments, {results}, "Visualize command positional arguments should match expectation.");
+
+            for f = ["FrequencyPoints", "Overlap", "Window"]
+                testCase.assertThat(command.NamedArguments, mag.test.constraint.IsField(f), compose("""%s"" should be a named argument.", f));
+            end
+
+            testCase.verifyEqual(command.NamedArguments.FrequencyPoints, 256, """FrequencyPoints"" should match expectation.");
+            testCase.verifyTrue(ismissing(command.NamedArguments.Overlap), """Overlap"" should match expectation.");
+            testCase.verifyTrue(ismissing(command.NamedArguments.Window), """Window"" should match expectation.");
+        end
+
+        % Test that "getVisualizeCommand" returns expected command, when
+        % "Overlap" is modified.
+        function getVisualizeCommand_modifiedOverlap(testCase)
+
+            % Set up.
+            panel = testCase.createTestPanel();
+
+            spectrogram = mag.app.control.Spectrogram(panel);
+            spectrogram.instantiate();
+
+            spectrogram.OverlapSpinner.Value = 0.5;
+
+            results = mag.Instrument();
+
+            % Exercise.
+            command = spectrogram.getVisualizeCommand(results);
+
+            % Verify.
+            testCase.verifyEqual(command.PositionalArguments, {results}, "Visualize command positional arguments should match expectation.");
+
+            for f = ["FrequencyPoints", "Overlap", "Window"]
+                testCase.assertThat(command.NamedArguments, mag.test.constraint.IsField(f), compose("""%s"" should be a named argument.", f));
+            end
+
+            testCase.verifyEqual(command.NamedArguments.FrequencyPoints, 256, """FrequencyPoints"" should match expectation.");
+            testCase.verifyEqual(command.NamedArguments.Overlap, 0.5, """Overlap"" should match expectation.");
+            testCase.verifyTrue(ismissing(command.NamedArguments.Window), """Window"" should match expectation.");
+        end
+
+        % Test that "getVisualizeCommand" returns expected command, when
+        % "Window" is modified.
+        function getVisualizeCommand_modifiedWindow(testCase)
+
+            % Set up.
+            panel = testCase.createTestPanel();
+
+            spectrogram = mag.app.control.Spectrogram(panel);
+            spectrogram.instantiate();
+
+            spectrogram.WindowSpinner.Value = 25;
+
+            results = mag.Instrument();
+
+            % Exercise.
+            command = spectrogram.getVisualizeCommand(results);
+
+            % Verify.
+            testCase.verifyEqual(command.PositionalArguments, {results}, "Visualize command positional arguments should match expectation.");
+
+            for f = ["FrequencyPoints", "Overlap", "Window"]
+                testCase.assertThat(command.NamedArguments, mag.test.constraint.IsField(f), compose("""%s"" should be a named argument.", f));
+            end
+
+            testCase.verifyEqual(command.NamedArguments.FrequencyPoints, 256, """FrequencyPoints"" should match expectation.");
+            testCase.verifyTrue(ismissing(command.NamedArguments.Overlap), """Overlap"" should match expectation.");
+            testCase.verifyEqual(command.NamedArguments.Window, 25, """Window"" should match expectation.");
+        end
     end
 end